pci: refactor: bring pci/mod.rs into vmm/pci
The pci crate had only one file with couple of types
only used in the vmm crate. And vmm crate already has
pci module as well, so move pci/mod.rs into vmm/pci/mod.rs.
