--- src/pci/Cargo.toml
@@ -1,24 +1,0 @@
-[package]
-authors = ["Samuel Ortiz <sameo@linux.intel.com>"]
-edition = "2021"
-name = "pci"
-version = "0.1.0"
-license = "Apache-2.0 AND BSD-3-Clause"
-
-[lib]
-bench = false
-
-[features]
-default = []
-
-[dependencies]
-
-displaydoc = "0.2.5"
-libc = "0.2.183"
-log = "0.4.29"
-serde = { version = "1.0.228", features = ["derive"] }
-thiserror = "2.0.18"
-
-[dev-dependencies]
-serde_test = "1.0.177"
-vmm-sys-util = "0.15.0"
